feat: rework on how client components recieve db data
