+++ app/ui/Home/Gallery/Gallery.scss
@@ -21,7 +21,7 @@
     position: absolute;
     bottom: -25px;
     left: 15px;
-    background-image: url('../../../public/icons/pointing.png');
+    background-image: url('../../../../public/icons/pointing.png');
     background-position: center;
     background-size: cover;
     background-repeat: no-repeat;
