+++ app/ui/Home/Gallery/Gallery.tsx
@@ -5,27 +5,15 @@ import "react-responsive-carousel/lib/styles/carousel.min.css";
 import { memo, useEffect, useState } from "react";
 import { Carousel } from "react-responsive-carousel";
 import { type GalleryType } from "@/app/lib/definitions";
-import { getGallery } from "@/app/lib/data";
 import SocialIcon from "@/app/ui/common/SocialIcon/SocialIcon";
 import Flickr from "@/public/icons/flickr.svg";
 import Instagram from "@/public/icons/instagram.svg";
 import Xcom from "@/public/icons/Xcom.svg";
-import Loading from "@/app/loading";
 
 
-const Gallery = memo(() => {
-  const [gallery, setGallery] = useState<GalleryType[]>();
+const Gallery = memo(({ gallery }: { gallery: GalleryType[] }) => {
 
-  useEffect(() => {
-    const fetchGallery = async () => {
-      const result = await getGallery();
-      setGallery(result);
-    };
-
-    fetchGallery();
-  }, []);
-
-  if(!gallery) return <Loading />;
+  if(!gallery) return <>emmm</>;
   
   return (
     <div className="card__content card__gallery animate-render">
